Moves Task to Job for security and fewer class files
