+++ src/applications/Job.java
@@ -50,4 +50,32 @@ class Job {
         return id;
     }
 
+    public Task getFirstTask(){
+        return (Task) taskQ.getFrontElement();
+    }
+
+    public int getTaskMachine(Task theTask){
+        return theTask.getMachine();
+    }
+
+    // top-level nested classes
+    private static class Task {
+        // data members
+        private int machine;
+        private int time;
+
+        // constructor
+        Task(int theMachine, int theTime) {
+            machine = theMachine;
+            time = theTime;
+        }
+
+        public int getMachine() {
+            return machine;
+        }
+
+        public int getTime() {
+            return time;
+        }
+    }
 }
