+++ src/applications/MachineShopSimulator.java
@@ -30,7 +30,7 @@ public class MachineShopSimulator {
             return false;
         } else {// theJob has a next task
                 // get machineArray for next task
-            int p = ((Task) theJob.getTaskQ().getFrontElement()).getMachine();
+            int p = theJob.getTaskMachine(theJob.getFirstTask());
             // put on machineArray p's wait queue
             machineArray[p].getJobQ().put(theJob);
             theJob.setArrivalTime(timeNow);
